Create 1 network service for V1 and other for V2
