+++ client/app/index.html
@@ -111,6 +111,8 @@
   <script src="src/services/config.service.js"></script>
   <script src="src/services/dialog.service.js"></script>
   <script src="src/services/storage.service.js"></script>
+  <script src="src/services/network-v1.service.js"></script>
+  <script src="src/services/network-v2.service.js"></script>
   <script src="src/services/network.service.js"></script>
   <script src="src/services/market.service.js"></script>
   <script src="src/services/feed.service.js"></script>
